Improve performance by removing extra GET request
Save one HTTP request, by updating the choices[] locally. If the POST
is succesfull then map through selected question choices, find the choice
that has changed based on string and update it. String comparison is
unsafe and should be avoided. One solution would be to ask backend to
provide id of the question as well, or we could extract it by the choice
url parameter by some string manipulation.